Overview example is now working. I am not very happy with how it works:
* There is a need to tune hyperparameters. Max hash size (this is fine).
The branch conditions in this example is pathological. So the system
ends up repeating a lot of work and backtracking for if then and repeats
it for the else branch. Guessing the else branch condition by putting !
method calls significantly speeds up this process. In this case the code
generated is fine, but often in other cases it is not. The code is
always correct though, it just doesn't look natural. So branch guessing
is now a hyperparameter too.
* I forgot what else I wanted to write.